Turned on description file generation for TopDesign.cysch
XML files are now generated that describe the binary file TopDesign.cysch. Changes to TopDesign should still be committed seperate of changes to source files, but the .cysch.cysem and .cysch.cyvis files should also be included in that commit. 
In this commit, no functional changes were made to TopDesign, only cosmetic to force generation of the description files.